Refactor: Split util module into separate files
Split src/cli/util.rs into:
- src/cli/util/mod.rs - module coordinator
- src/cli/util/trailing_whitespace.rs - trailing whitespace functionality

This makes the code more modular and easier to extend with new utility
commands.

All existing tests pass.
