+++ src/cli/util/mod.rs
@@ -1,0 +1,26 @@
+mod trailing_whitespace;
+
+pub use trailing_whitespace::TrailingWhitespace;
+
+use crate::Result;
+
+/// Utility commands for file operations
+#[derive(Debug, clap::Args)]
+pub struct Util {
+    #[clap(subcommand)]
+    command: UtilCommands,
+}
+
+#[derive(Debug, clap::Subcommand)]
+enum UtilCommands {
+    /// Check for and optionally fix trailing whitespace
+    TrailingWhitespace(TrailingWhitespace),
+}
+
+impl Util {
+    pub async fn run(&self) -> Result<()> {
+        match &self.command {
+            UtilCommands::TrailingWhitespace(cmd) => cmd.run().await,
+        }
+    }
+}
